Propagate environment variables to run and test unless `def propagateEnv = false` (#5617)
Fixes https://github.com/com-lihaoyi/mill/issues/5430, and would be
necessary to make `UTEST_UPDATE_GOLDEN_TESTS` work smoothly

Previously, we relied on the default `propagateEnv = true`, but that
doesn't work well because it picks up the Mill daemon process's env vars
which do not get updated when you re-run the launcher. This PR
explicitly sets `propagateEnv = false`, and passes in the `Task.env`
explicitly to `Jvm.callProcess` when the `RunModule#propagateEnv` task
is configured to be true (the default). If `RunModule#propagateEnv` is
`false`, then we do not propagate any environment variables from the
Mill process to the subprocess

Covered by a new `integration/feature/env-vars/` test that fails on main
and passes on this PR