Fix android install (#5623)
Fixes caches not being invalidated for androidDex when source code
changes

ref: https://github.com/com-lihaoyi/mill/issues/5616

## Explanation

- androidD8Dex was just returning the cli arguments with the paths as
strings, thus while it was executed, androidDex was not, because the
result from `androidD8Dex` was identical even after source changes.
After adding the list of PathRef of the compiled classes the androidDex
was executed on altering source files as in turn, compiled files were
changed and detection is propagated by path ref hashing.
- androidDex was forwarding the PathRef of `androidD8Dex` which was
being created on an empty directory, thus the PathRef signature was
identical. Recreating it gives a fresh hash.

## Fix demo


https://github.com/user-attachments/assets/76a85669-c62b-43a7-85e4-5931f2f53b6d