Added new implementation for SBT importer with support for cross-platform and cross version projects
Cross-platform support is limited to projects using `sbt-crossproject` plugin and was achieved with custom layout supertypes that replicate the project structure used by the plugin.

Project settings are imported by running a custom script in the SBT shell. This means that the values imported are for the default Scala version used in the project. Post import, running Mill tasks with a different version may result in failures. For now, a warning message is logged notifying the user of this possibility.