Fix race condition in JsonArrayLogger shutdown (#5636)
Using `.interrupt` on the thread after waiting for `buffer` to be empty
is risky, because it can interrupt the thread in a whole bunch of
difference places. Better to pass in a sentinel value to make sure the
thread shuts down gracefully

Tested manually using `./mill-assembly.jar core.__.test`. Previously it
would sometimes hang at the end of the test run, with this PR I haven't
managed to make it hang. I'm not fully confident in the exact mechanism
of the original hang, but at least it seems to have gone away

Relevant stack trace of the hang:

```
"main" #1 prio=5 os_prio=31 cpu=1213.49ms elapsed=37.88s tid=0x000000013181ac00 nid=0x2803 waiting on condition  [0x000000016deac000]
   java.lang.Thread.State: TIMED_WAITING (sleeping)
	at java.lang.Thread.sleep(java.base@17.0.14/Native Method)
	at mill.internal.JsonArrayLogger.close(JsonArrayLogger.scala:62)
	- locked <0x0000000601dd8030> (a mill.internal.JsonArrayLogger$Profile)
	at mill.exec.Execution.close(Execution.scala:323)
	at mill.eval.EvaluatorImpl.close(EvaluatorImpl.scala:271)
	at mill.testkit.UnitTester.closeWithoutCheckingLeaks(UnitTester.scala:236)
	at mill.testkit.UnitTester.close(UnitTester.scala:240)
	at mill.testkit.UnitTester.scoped(UnitTester.scala:229)
	at mill.exec.ModuleTests$.$init$$$anonfun$1$$anonfun$2(ModuleTests.scala:52)
	at mill.exec.ModuleTests$$$Lambda$640/0x000000b80138c2c8.apply(Unknown Source)
	at utest.framework.StackMarker$.dropOutside(StackMarker.scala:13)
```