Tab-complete `--flag`s belonging to `Task.Command`s as well  (#5643)
Rather than treating everything past `leftoverArgs.head` as unknown and
delegating to bash, we first check if `leftoverArgs.head` is a command.
If it is, then we can load it's `mainargs.MainData`, and use that to try
and parse the remaining tokens. From there we can try and identify
half-completed `--flag` tokens at the current index, and autocomplete
them

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>