Fix flicker in PromptLogger reporting of current test class running for each worker (#5675)
Previously every iteration we would loop through the entire `claim.log`
file and emit a `logger.ticker` for every class found. Depending on
scheduling could result in older already-complete classes being shown in
the prompt for a fraction of a second before being over-written, which
would cause a noticeable flicker in the prompt lines. This PR makes us
explicitly
track the last test class and it's start `nanoTime`, so we don't
`logger.ticker` earlier already-completed classes that are no longer
relevant

Tested manually on our example Netty build's `common.test`; the flicker
is noticeable on `1.0.3`, goes away with this patch