Android: Proguard issues fixes (#5677)
Fixes an issue of Proguard usage as mentioned in
https://github.com/com-lihaoyi/mill/issues/5654.

> Proguard rules for Hilt need to be autogenerated along with
initialisers in the android manifest provider section to avoid R8
shrinking "nullifying" the constructors

#### EDIT:
This PR does not resolve the cache invalidation issue after all 
> Proguard file changes do not invalidate androidDex/androidApk caches  

The temporary fix we initially considered would have required API
changes, which are not desired.
We will instead follow up with a PR that fixes the issue without
altering the APIs.