Draft: Use custom JVM versions during Scala compilation, avoid forking for Java compilation (#5604)
Adds support for compiling projects under different JVMs and JVM runtime
options.

Previously:
- Compiling java sources with a different JVM/options worked, but a
compiler was forked for each compilation, which meant that it was always
JIT-cold.
- Specifying `jvmId` in a `ScalaModule` did not make the compiler run
with that JVM. The option was ignored.

This PR fixes both of those issues.

To do that, it introduces an RPC framework to mill, located under
`mill.rpc` namespace.

### RPC overview

`MillRpcClient.create` allows you to create an RPC client using a
specific `MillRpcWireTransport`. A transport is a simple trait that
provides a way to read and write protocol messages to the wire. RPC uses
line-delimited JSON for exchanging messages.

Currently supported transports are:
- Using stdin/stdout.
- Using stdin/stdout of a subprocess.
- Using concurrent blocking queues (mostly used for testing).
- Using Java streams (for example, from a TCP connection).

Client, upon establishing a connection to the server, sends an
`Initialize` message. Then it sends `ClientToServer` messages, which
always return a response (which can be `Unit`). A server, when
responding to a client's message, can also send a `ServerToClient`
request to the client (for example, to obtain additional data from the
client needed to fulfill the request). However, in the current
implementation, a server cannot push messages to the client without the
client initiating a request first (as the client implementation does not
read from the transport unless it is waiting for a server's response).

### TCP server infrastructure

Part of the code has been extracted from `runner.client` and
`runner.server` modules into `libs.daemon.client` and
`libs.daemon.server` respectively.

This code handles:
- Client starting a server process if needed, or connecting to a running
process if it's there.
- Initial handshake between the client and the server via the filesystem
to exchange on which free port the server did bind.
- Server accepting connections on separate threads and the client
establishing those connections.
- Multiplexing stdout/stderr from the server into a single TCP stream
and demultiplexing it on the client side.
- Infrastructure for returning exit codes to the client and other
conveniences.

### ZincWorker

Keeps the in-memory Java/Scala compilers and performs the actual
compilations. There is one local `ZincWorker` in `JvmWorkerImpl` which
is used if the compilation is performed without a custom JVM and custom
JVM runtime options, and each `ZincWorkerRpcServer` also contains a
`ZincWorker`.

### ZincWorkerRpcServer

This handles requests to perform Java/Scala compilations and Scaladoc
generation. It uses `ZincWorker` internally to do that.

### ZincWorkerMain

Entry point for the TCP-based RPC server. `JvmWorkerImpl` spins up these
as JVM subprocesses and then communicates with them via TCP to perform
java/scala compilations and scaladoc generation.

Supports multiple clients, each client is handled in it's own thread.

--- 

Fixes https://github.com/com-lihaoyi/mill/issues/3982

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>