BSP fixes (#5698)
This fixes several small BSP issues, and tweaks the BSP snapshot tests
too.

In particular, this makes the BSP snapshot tests more exhaustive, by
testing a meta-build too, and making some tests more broad (main class
request test).

This fixes issues with the buildTargetInverseSources request:
- it makes return the correct target when passed a `build.mill` file
(while it doesn't return anything without the changes here)
- it ensures a compile reporter is around when evaluating tasks for that
request, which is required if a `sources` task depends on a `compile`
one for example (as a check requires such a reporter to be around during
all task evaluations in the BSP server, so that we don't miss any
diagnostic)

This PR commits can be reviewed one-by-one, and provide a bit of
explanation of what each commit does in the commit message.