KSP 2 support (#5727)
## Motivation: 

https://github.com/com-lihaoyi/mill/issues/5676

Also, language version 1.9 is deprecated since Kotlin 2.2.0. 

It turns out, that integration wise, it's much easier to run the symbol
processor with KSP 2 via a CLI wrapper that it is for KSP 1 as a kotlin
compiler plugin


## Details

The documentation for the CLI wrapper is
[here](https://github.com/google/ksp/blob/main/docs/ksp2cmdline.md)
albeit a bit outdated. The key component that is missing, is a reference
to the `-libraries` argument, which (as our trial and error shows)
expects the module classpath.

Micronaut had also a tricky bit, failing to resolve java sdk sources.
Passing the `-jdk-home` resolves that.

From there on, Dagger, Hilt and micronaut work without any major
adaptations from the user's api perspective as we made an effort to
consolidate `KSPModule` and `KSP2Module` input parameters

We have also retested the pokemon module outside of this repo which also
works, surprisingly without needing the `-friends` parameter

## Summary

This PR gives mill the capability to build KSP 2 with language version
and api version >= 2.0 , bringing a solution to the deprecated language
version <=1.9 needed by KSP 1.

Also, added an equivalent Micronaut example as the hello-micronaut in
javalib/web, but with full KSP, no annotation processors needed

## Credits

This PR was co-authored by @souvlakias

---------

Co-authored-by: souvlakias <stylianos@vaslabs.io>