Rework Android SDK installation
This refactored the task dependency flow of `AndroidSdkModule`. Previously, most task provided `PathRef`s to tools depend on some anonymous installation tasks which need to synchronize their work.

Now, the installation steps are bound to single tasks, which returns a handle to the installed resources, which are:

* `cmdlineTools`
* `androidSdk`
* `androidNdk`

Existing task pointing to various resources and executables, now just depend on these installation tasks and simple forward the returned resource.