Rework Android SDK installation (#5766)
This refactors the task dependency flow of `AndroidSdkModule`.
Previously, most tasks provided `PathRef`s to tools depending itself on
some anonymous installation tasks which needed to synchronize their
work.

Now, the installation steps are bound to single tasks, which return a
handle to the installed resources. These tasks are:

* `cmdlineTools`
* `androidSdk`
* `androidNdk`

Existing tasks pointing to various resources and executables, now just
depend on these installation tasks and simply forward the returned
resources.

Additional on Windows, tools will be searched for with the extensions
`exe` and `bat`.

This is not the final design, but a step forward in making the setup
more idiomatic and readable.

Pull request: https://github.com/com-lihaoyi/mill/pull/5766