Let `MillURLClassLoader` find classes not found in shared classloader (#5790)
This fixes an issue with our current `MillURLClassLoader` implementation
regarding package prefixes. The issue is, that if a users package is a
sub-package of one of the pre-configured shared package prefixes,
classes will only be searched in the shared classloader, but not in the
`MillURLClassLoader`. This leads to `ClassNotFoundError`s although the
class is present on the classpath represented as URLs. In one specific
case, Scalameta, which uses the `scala.meta` package (a sub-package of
`scala` which we have in our pre-configured package prefix list), could
not be used in a users project.

This is a trade-of between usability and correctness. Ideally, we would
not match sub-packages but exact packages, but it's likely we miss some
or run into compatibility and maintenance issues whenever a single part
changes, e.g. Java version or Scala versions. It might even be not
exactly configurable at all given the history of packages like `javax`.

Instead, maintaining a fixed positive list of exceptions is equally bad,
since it needs to be maintained as well and will most likely result in
bad user experiences, esp. in environments, where closed or non-open
source libraries are used.

The solution of this PR does not require any special maintenance, but
has the potential to give misleading error messages, since we won't
detect classloader misconfigurations as exactly as before. While I don't
expect it being an issue often, it may still happen, especially, when
developing or changing more fundamental parts of Mill like the
classloader separation and isolation setup. (But it is very easy to
disable this feature while refactor, so this isn't such a dramatic
downside.)

Added an integration test under `integration/feature/mill-classloader`.

Fix https://github.com/com-lihaoyi/mill/issues/5759

Pull request: https://github.com/com-lihaoyi/mill/pull/5790