Try removing `semanticDbData` Java compiler flags (#5798)
Attempts to fix https://github.com/com-lihaoyi/mill/issues/5741

Kicking this off to see if anything breaks in the test suite. 

If nothing breaks, we can rebootstrap on top and see if anything breaks.

If nothing breaks rebootstrapping, it's probably fine since that covers
building a huge pile of Java/Scala code on Java 11, 17, and 24

If something breaks in user code when we release it, we can revert the
change and add a test case

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>