Fix forced versions (#5813)
This fixes https://github.com/com-lihaoyi/mill/issues/5796 (forced
versions not enforced) by taking them into account in
`JavaModule#coursierDependency`, so that they're being enforced during
dependency resolution, and are taken into account in other modules if
needed.

In order to get forced versions from `mvnDeps`,
`JavaModule#coursierDependency` needs to be a task. So in more detail,
the current `coursierDependency` is deprecated, and a new
`coursierDependencyTask` is used instead.

---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>