Update lazy val bytecode signature heuristics to work with Scala 3 lazy vals (#5828)
Should fix https://github.com/com-lihaoyi/mill/issues/5356

The previous heuristics worked for Scala 2, but weren't covered by
tests, and so bitrotted in Scala 3. This PR updates the heuristics to
the bytecode I can see being emitted by the Scala 3 compiler, and adds
an integration test assertion to make sure that if it regresses again we
can notice. The existing tests should hopefully catch any regression in
the codesig robustness