Retry mill daemon start rather than just PID file reading (#5829)
Fixes https://github.com/com-lihaoyi/mill/issues/5800, hopefully

My understanding of the problem is that sometimes the launcher connects
to a daemon in the process of shutting down; this is most commonly the
case when changing Mill versions, and the shutdown is immediately
followed by a restart, so the previous daemon may not have fully exited
yet.

In such cases, the previous logic may see the `daemonLock` still taken,
assume the daemon is alive, and start retrying reading the PID file. But
if the daemon was in the process of shutting down, the PID file may
never exist, and polling eventually fails

The new logic broadens the retries to also include the
`locks.daemonLock.probe()` check, such that on retry the launcher has a
chance to notice if the daemon that was previously around is now gone,
and it can then take actual to start a new daemon.

Not sure how to test this, since even the previous misbehavior is
nondeterministic and I haven't managed to reliably reproduce it

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>