Fix prompt when server restarts (#5832)
Fixes https://github.com/com-lihaoyi/mill/issues/5637

The basic problem was that `prepareMillRunFolder` gets called via
`MillServerLauncher#run` and `prepareDaemonDir`, which can end up
calling it twice when the daemon restarts since we call `.run` twice in
`MillLauncherMain`. But `memoizedTerminalDims` gets set the first time
and not cleared, and so the second time `prepareMillRunFolder` gets run
it deletes the old `terminfo` file but thinks the dimensions are
unchanged (v.s. the `memoizedTerminalDims`) and so does not write a new
one

The fix is to move `prepareMillRunFolder` into
`configureRunMillProcess`, so it only gets called when we actually spawn
the Mill daemon process, which should only happen once per launcher.
This also lets us remove `prepareDaemonDir` and simplify things overall