`--bsp` should imply `--no-daemon` (#5840)
Ensure that if we're running in BSP mode we don't start a daemon.

This is needed because when Metals/Idea closes, they only kill the BSP
client and the BSP server lurks around waiting for the next client to
connect.

This is unintuitive from the user's perspective and wastes resources, as
most people expect everything related to the BSP server to be killed
when closing the editor.

Fixes https://github.com/com-lihaoyi/mill/issues/5839

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>