Clean up daemon/server launching retry logic to try and stabilize it (#5865)
Previously the retries of each step (1) ensure server alive (2) read
port (3) connect to socket, happened sequentially with no opportunity
for backtracking, which could cause problems if the server was in a bad
state, e.g. if the server was alive but shutting down:

- (1) may succeed but then the server shuts down, deletes the port file,
and (2) may never become possible
- (1) may succeed and (2) may succeed but then the server shuts down and
(3) may then never work

This PR changes the logic such that the retries for (1) and (2, 3) are
nested: we retry the server-alive check, if that succeeds we try to read
the port and connect to it, but if either of those fails we go back to
retry the server alive check again. This ensures that in the server
shutting down scenario, we get a chance to launch a new server if port
reading or socket connection fails

Tested by alternating `./mill-assembly.jar version` and `./mill
version`. With an assembly built in `main`, it sometimes fails with
`Failed to connect to server`. With an assembly built on this PR, I
haven't managed to get it to fail.

Also tweaked the `JavaHomeTests` to try and improve performance