manage BOMs in Tests
Starting with JUnit 5.12 and Spock 2.3 a BOM is available on Maven Central.
If the version matches the BOMs will be automatically added.

For JUnit, this has the effect that the platform-launcher is also managed in the BOM so it is actually not necessary to specify a platform-launcher-version. Hence the launcher is also added when the version is not set and the minimum version is met.