Blog Post: How does mill build android apps (#5776)
An introductory blog post for Mill's android support. I tried to focus
on celebrating mill's design in helping to add all this android support.
I think I need a better title, please let me know your ideas and also
feedback for the content.

Also added a page in the android docs for the native endless tunnel
example (so as I can also reference it in the blog post).

---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>