Fix prompt resizing when connecting to existing daemon (#5924)
Fixes https://github.com/com-lihaoyi/mill/issues/5919

This regressed in https://github.com/com-lihaoyi/mill/pull/5832, which
made us only start the `terminfo` writing thread when spawning a new
server, which neglects the case where we are connecting to an existing
server. Thus when you ran `./mill clean && ./mill __.compile`, the
`__.compile` runs with a launcher connected to the existing server, and
thus doesn't update when the terminal size changes

This PR makes us start the `terminfo` thread regardless of whether we
are spawning a new server or not. Tested manually

Also did some cleanup of `libs/daemon/client` to remove dead code.