Fix #4584 Scala.js OOM when running `fullLinkJS`
We use `mill.util.CachedFactoryWithInitData` with a limit of jobs entries.
This way we never store more than `jobs` number of working linkers.
Moreover we split the cache in two separate caches:
- `fastLinkJS` has maxCacheSize of `jobs` so we can have multiple warmed up
   scalajs modules linked
- `fullLinkJS` has maxCacheSize of 1, since we don't want to consume too
   much memory. If you need to run `fullLinkJS` on two modules the second
   linker will run only after cleaning up the cache of the previous Linker.
Now `ScalaJSWorkerApi` extends `AutoCloseable`, so when we close the
Classloader, before that we also clean all the Scala.js linker caches.