feat: support Playwright testing for Scala-JS (#5777)
I could not find any existing JSEnv which allows me to properly test
against some browser features and am using
[scala-js-env-playwright](https://github.com/ThijsBroersen/scala-js-env-playwright)
for a while now. The current version is still a snapshot so this PR is
not ready to be merged, but I would love some feedback on what is
required to get this feature as part of Mill. Currently I am using a
custom build mill-assembly for my projects.

Here a draft of Playwright support for Scala-js testing. 

I was also wondering why isn't Mill supporting just a raw JSEnv? Any
custom JSEnv implemention now first requires Mill to include it..

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>