Update remaining examples still referencing `JvmWorkerModule.jvmId` (#5950)
With the move of `jvmId` from `JvmWorkerModule` to `JavaModule` in PR
#5327 a few examples were not updated.

<!--
Please open all PRs as drafts and ensure that your fork of Mill has 
`settings/actions` / `Allow all actions and reusable workflows` enabled
to run CI on
your own fork of the Mill repo. Only once CI passes mark the PR as
`Ready for review`
and CI will run on the main Mill repo before we merge it.
-->