Re-enable `--repl` (#5983)
This PR resurrects the `--repl` command to open up a meta-build
`console`. This allows Mill to provide a decent "default" Scala REPL
with the bundled libraries available (os-lib, upickle, requests,
pprint), which together with the recent improvements in the upstream
Scala REPL should make `./mill --repl` a viable alternative to Ammonite.

We explicitly choose to open a meta-build `console` rather than a normal
`console` to avoid collision with any `console` command that may be
present in the user's root module: since the meta-build must be a
`MillBuildRootModule`, we can be certain it is a `ScalaModule` and also
it will have the relevant libraries on the classpath (along with some
others). It needs to be a flag rather than a command so we can parse it
in `MillLaunchrMain` and decide to run in no-daemon mode, and also so we
can parse it early enough to set `--meta-level` to `1` to implement the
semantics above

For now we don't add any helpers to assist in evaluating tasks, but that
can be done in a follow up if necessary

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>