Include contrib modules in API docs and improve `scalapblib` docs (#6001)
We now include all contrib modules in our API documentation. This
removes the need to document the available configuration task in two
places, the code and the readme files. For convenience, contrib readmes
should point to their API docs. I already did this for `scalapblib`, to
test it out. If it works reasonable, we can rework the other plugins as
well.

Pull request: https://github.com/com-lihaoyi/mill/pull/6001