Fix Scala Native link logs in server-client mode
The Scala Native toolchain uses its own threads to print using the logger. System.err on Scala Native threads doesn't have the Mill setup to forward the output to the client.
So we store Console.err at the beginning of the task (in the Mill thread) which has the correct setup and use that reference to log.
This way logs always reach the terminal correctly.
Port for Scala Native of the following Scala.js related Pull Request: https://github.com/com-lihaoyi/mill/pull/5845