Allow users to disable some modules in BSP (#6065)
This adds a `def enableBsp: Boolean` method on `JavaModule` (via
`BspModuleApi`). Users can override this method and make it return
`false` to make the BSP server hide / ignore that module.

Note that for consistency, if a module depends on another that has
`enableBsp` set to false, it gets itself disabled too. Else, that might
create situations where a module is loaded in BSP, but some of its
dependencies would just be missing because they have `enableBsp` set to
false.