Run examples with local repo under out/ rather than ~/.ivy2/local (#6070)
This changes how Mill is run in integration tests and examples. This
makes it stop publishing things under `~/.ivy2/local`, and use something
like `out/dist/localRepo.dest` instead, which is the `Task.dest` of the
new `dist.localRepo` task. That new repository is a local Maven
repositiory rather than an Ivy one, which has a slightly different file
structure, the same as Maven Central.

For that new repository to be taken into account, the Mill launcher used
in examples and ITs is now a script (bat under Windows, bash everywhere
else) that sets the `COURSIER_REPOSITORIES` env var, so that the new
local repo is added to it, and gets automatically used when running
examples and ITs.

The second commit of this PR mainly updates the file names of the
locally published Mill artifacts, because of the switch from an Ivy
repository to a Maven one.

That's been useful to me when using a locally published Mill version, to
develop on Mill or other projects. Running examples or ITs erases it,
which is annoying, but also somewhat unexpected (running tests shouldn't
erase things in the local repository, at least not things put there
manually on purpose). This PR fixes that.

---

This adds back https://github.com/com-lihaoyi/mill/pull/5949 that was
reverted in https://github.com/com-lihaoyi/mill/pull/6067, and changes
`dist.installLocalCache`, so that it publishes things under
`~/.ivy2/local` like `dist.installLocal` does (rather than using the
local repo under `out/`, which is more likely to go away).