Fixes #6138: HtmlScraper is a flakey test. #6138
The original code goes out to Wikipedia's singapore page. From there it would print out all of the links that it found and then verify the output. The way that HtmlScraper prints out this gets truncated and causes the test to fail. The contents on the page put the link of the Government Of Singapore beyond the amount that was printed on pprint. This also changes the source of what's being parsed to a local file. (Making this test more stable)