Make selective execution trigger on changes to build overrides (#6158)
Build overrides need to be treated similarly to code signatures, and
this PR does that. This allows changes to build overrides to trigger
selective execution just as code changes can trigger it

Tested manually via:

- `./mill -i dist.run example/javalib/basic/1-script -iw
Foo.java:resolvedMvnDeps`
- `./mill -i dist.run example/javalib/basic/3-simple -iw
resolvedMvnDeps`
- `./mill -i dist.run example/extending/imports/1-mvn-deps -iw
--meta-level 1 resolvedMvnDeps`

And verified that while the watch was taking place, changing one of the
`mvnDeps` to an invalid string would cause `resolveMvnDeps` to re-run
and fail. Previously it would not re-run until the watch was cancelled
and restarted (which skips selective execution)