Move build-override handling back into evaluator (#6167)
Compared to having them scoped to the classloader in
`build-overrides.json` and loaded for each module via `def
moduleBuildOverrides`, having them scoped to the evaluator allows
handling to be more fine-grained. In particular, this means changes to
build overrides just need a new `Evaluator` to be created (which we do
every time anyway), but no longer needs the `build.mill` classloader to
be discarded and re-created to allow it to pick up updated
`build-overrides.json` resource files.

This means we can preserve most of the state in the classloader,
including most instantiated modules, and importantly heavyweight worker
tasks like the `ZincWorker` classloaders or subprocesses. That in turn
should reduce turnaround time when updating a `build.mill.yaml`,
`package.mill.yaml`, or script build header.

One compromise needed to support this is that the `ScriptModuleInit`
cache needs to be re-created every time, and all `ScriptModule`s
discarded and re-instantiated. This is because now that we are not
throwing away the classloader every time a `build-override.json` file
changes, it has revealed issues in the `ScriptModuleInit` cache
management and eviction logic that seem practically unfixable (e.g.
modules keeping references to stale already-evicted modules, either
directly or indirectly through task inputs)