Fix parsing/rendering of compile errors when the code snippet contains a `|` (#6198)
The previous renderer was incorrectly stripping the leading `|` twice,
which was incorrect when the code snippet itself contained a `|`.

Tested manually by forcing a compile error on a code line with a `|` in
it. Previously the renderer would strip everything before the code's
`|`, with this PR the renderer would only strip the line-number-gutter's
`|`

Also tweaked the rendering of paths to un-color the `/` and `\`s so
they're easier to scan