Make compilation operation work dir explicit
For one, this makes the code more readable. Being internal code, there is no reason to work with an implicit context here.

Also, it's much easier to use the code outside of a task-context, e.g. in case we need multiple compilation scopes in the same task, but with different work directories.