Add PublishCredentialsModule, MavenPublish and MavenPublishModule (#6023)
### Summary

Implements #5592 and builds upon #5601.

### High Level Overview of Changes

This PR implements three traits:

- `MavenPublishModule` is meant to mirror `SonatypeCentralPublishModule`
for plain (!?) Maven repositories
- `MavenModule` contains the actual publishing logic. This has been
taken from parts of `SonatypeCentralPublishModule` and is now extended
by both `MavenPublishModule` as well ass `SonatypeCentralPublishModule`
to share code
- `MavenCredentialsModule` contains tasks for obtaining Maven/Sonatype
credentials. These tasks have been extracted from the
`SonatypeCentralPublishModule` to be shared by both it and the new
`MavenPublishModule`

### Differences between this PR and #5601 

#### Code Structure

#5601 essentially copied parts of `SonatypeCentralPublishModule` to
create `MavenPublishModule`. Apart from duplicating code, this also
created code you may find confusing in places (e.g. by containing a
function named
[publishSnapshot](https://github.com/com-lihaoyi/mill/pull/5601/files#diff-7ffbad225d6fa5d1de00e7ea2066f029c9fe4b90c09785ac3cd51ea0d12b3638R126)
which is used to publish both release and snapshot artifacts). This PR
tries to improve upon that by instead factoring out all code which can
be factored out.

#### Removal of Unused Settings

#5601 defines the tasks `mavenConnectTimeout`, `mavenReadTimeout` and
`mavenAwaitTimeout` as part of the `MavenPublishModule`. These are
however not actually used anywhere. They are used in
`SonatypeCentralPublishModule` and are specific to publishing releases
to Sonatype Central.

#### Bugfix for Choosing the Repo URI

The below was changed after being identified during testing with a
private build.

```diff
- val uri = if (isSnapshot) releaseUri else snapshotUri
+ val uri = if (isSnapshot) snapshotUri else releaseUri
```

### Open Question: Build Compilation Error When Extending
`MavenPublishModule`

When running `./mill dist.installLocalCache` and then using the local
`SNAPSHOT` version of mill including the changes made in this PR in a
build and extending `MavenPublishModule` in a local build module, I
currently get the following exception when running `./mill resolve _` on
that build:

```
[build.mill-59] [error] ## Exception when compiling 23 sources to /home/max/Repositories/github.com/ttmzero/rtp-backend/out/mill-build/compile.dest/classes
[build.mill-59] [error] dotty.tools.dotc.core.MissingType: Cannot resolve reference to type com.lumidion.sonatype.central.client.core.type.SonatypeCredentials.
[build.mill-59] [error] The classfile defining the type might be missing from the classpath.
```

This error can be removed by defining `Deps.sonatypeCentralClient` as a
`mvnDeps` instead of both a `compileMvnDeps` and a `runMvnDeps` in
`libs/scalalib/package.mill` and `libs/javalib/package.mill`. I do not
know why this is the case and would require some help here.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>